+++ src/swap/core/getParamsForToken.ts
@@ -1,3 +1,4 @@
+import { formatDecimals } from './formatDecimals';
 import type { GetQuoteParams, GetQuoteAPIParams } from '../types';
 
 /**
@@ -14,18 +15,4 @@ export function getParamsForToken(params: GetQuoteParams): GetQuoteAPIParams {
     amount: amountInDecimals ? amount : formatDecimals(amount, false, decimals),
     amountReference: amountReference || 'from',
   };
-}
-
-/**
- * Formats an amount according to the decimals. Defaults to 18 decimals for ERC-20s.
- */
-export function formatDecimals(
-  amount: string,
-  inputInDecimals: boolean = true,
-  decimals: number = 18,
-): string {
-  if (inputInDecimals) {
-    return (Number(amount) / 10 ** decimals).toString();
-  }
-  return (Number(amount) * 10 ** decimals).toString();
 }
